requirements-serv.txt
psycopg2-binary==2.8.6
redis
python-redis-lock==3.2.0
django==2.2
djangorestframework==3.11.0
django-filter==2.2.0
django-cors-headers==3.2.0
django-celery-beat==2.1.0
celery[redis]==5.1.2
channels==2.1.5
channels_redis==2.3.1
gunicorn==19.9.0
uvicorn==0.13.4
django-redis==4.11.0
django-bulk-update==0.1
drf-yasg==1.17.0
web3==5.0
bitcoin==1.1.42
ecdsa==0.13
sortedcontainers==2.1.0
